Added TxOut modules for each era
